fix(youtube-proxy): merge fixed variant into canonical proxy
V019.4 Resolution: Integrated youtube_proxy_fixed.py into canonical
youtube_proxy.py and deleted the redundant file.

Integrated methods:
- find_active_livestream() - Active livestream discovery with self-healing auth
- _auto_refresh_tokens() - WSP 48 automatic token refresh

Tests: 26/26 passing
