+++ WSP_framework/src/WSP_MODULE_VIOLATIONS.md
@@ -362,11 +362,12 @@ mv test_menu_input.txt tests/test_data/
 - **WSP 62 Size**: 317 lines - [U+2705] WSP compliant
 - **Action**: Module consolidated and compliant. Test imports updated.
 
-### **V019.4: YouTube Proxy Duplicates (platform_integration domain) - STATUS CONFIRMED**
+### **V019.4: YouTube Proxy Duplicates (platform_integration domain) - [U+2705] RESOLVED**
 - **Canonical**: `modules/platform_integration/youtube_proxy/src/youtube_proxy.py`
-- **Additional**: `modules/platform_integration/youtube_proxy/src/youtube_proxy_fixed.py`
-- **Finding**: Fixed variant present; treat as patch layer pending integration.
-- **Action**: Analyze patch deltas -> integrate into canonical -> retire fixed after tests.
+- **Integrated**: `find_active_livestream()` and `_auto_refresh_tokens()` methods
+- **Deleted**: `youtube_proxy_fixed.py` (redundant after integration)
+- **Resolution Date**: 2026-04-27
+- **Action**: COMPLETE - Fixed variant merged into canonical, tests pass.
 
 ### **V019.5: Stream Resolver Multi-Version Pattern (platform_integration domain)**
 - **Canonical**: `modules/platform_integration/stream_resolver/src/stream_resolver.py` (v0.1.5 locked)
@@ -384,7 +385,6 @@ mv test_menu_input.txt tests/test_data/
 
 **P1 - High (Next Development Session)**:
 - V019.1: Banter engine feature consolidation
-- V019.4: YouTube proxy patch integration
 
 **P2 - Documentation**:
 - V019.5: Stream resolver multi-version pattern documentation
