+++ modules/platform_integration/youtube_proxy/ModLog.md
@@ -2,6 +2,17 @@
 
 ## Latest Changes
 
+### **V019.4 Resolution: Fixed Variant Integration**
+- **Status**: [OK] RESOLVED
+- **Date**: 2026-04-27
+- **Change**: Merged `youtube_proxy_fixed.py` into canonical `youtube_proxy.py`
+- **Integrated Methods**:
+  - `find_active_livestream()` - Active livestream discovery with self-healing auth
+  - `_auto_refresh_tokens()` - WSP 48 automatic token refresh
+- **Deleted**: `src/youtube_proxy_fixed.py` (redundant after integration)
+- **Tests**: 26/26 passing
+- **WSP Compliance**: V019.4 module duplication violation resolved
+
 ### **WSP Hygiene: Root Script Relocation**
 - **Status**: [OK] Completed
 - **Change**: Moved YouTube diagnostic/engagement utilities out of repository root into `modules/platform_integration/youtube_proxy/scripts/manual_tools/` to satisfy WSP 3/85 (no vibecoding at root).
