+++ modules/platform_integration/youtube_proxy/tests/TestModLog.md
@@ -9,13 +9,14 @@
 
 ### **Testing Framework Established** [OK]
 
-### [2025-08-10 12:02:47] - YouTube Proxy Fixed Tests
-**Test Coverage**: Pending
-**Status**: ⏳ Tests needed for youtube_proxy_fixed.py
+### [2026-04-27] - V019.4 Integration Complete
+**Status**: [OK] `youtube_proxy_fixed.py` merged into canonical `youtube_proxy.py`
+**Methods Integrated**: `find_active_livestream()`, `_auto_refresh_tokens()`
+**Tests**: 26/26 passing (coverage via `test_youtube_proxy.py`)
 
-#### Required Tests
-- [ ] Test find_active_livestream method
-- [ ] Test self-healing authentication
+#### Remaining Test Coverage Gaps
+- [ ] Test find_active_livestream method with mocked stream_resolver
+- [ ] Test self-healing authentication flow
 - [ ] Test credential rotation on failure
 - [ ] Test fallback to stream_resolver
